+++ lib/design/widgets/appBar/back_appBar.dart
@@ -1,0 +1,29 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_svg/svg.dart';
+import 'package:get/get.dart';
+
+
+class BackAppBar extends StatelessWidget {
+  final Color iconColor;
+
+  const BackAppBar({
+    super.key,
+    required this.iconColor,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return AppBar(
+      backgroundColor: Colors.transparent,
+      leading: IconButton(
+        onPressed: () {
+          Get.back();
+        },
+        icon: SvgPicture.asset(
+          'assets/icons/back_left.svg',
+          colorFilter: ColorFilter.mode(iconColor, BlendMode.srcIn),
+        ),
+      ),
+    );
+  }
+}
